rename option `virtualenvs.prefer-active-python` and move whole logic to EnvManager.create_venv()
