+++ docs/configuration.md
@@ -48,7 +48,7 @@ virtualenvs.in-project = null
 virtualenvs.options.always-copy = true
 virtualenvs.options.system-site-packages = false
 virtualenvs.path = "{cache-dir}/virtualenvs"  # /path/to/cache/directory/virtualenvs
-virtualenvs.prefer-shell-python = false
+virtualenvs.prefer-active-python = false
 ```
 
 ## Displaying a single configuration setting
@@ -189,7 +189,7 @@ Give the virtual environment access to the system site-packages directory.
 Applies on virtualenv creation.
 Defaults to `false`.
 
-### `virtualenvs.prefer-shell-python`
+### `virtualenvs.prefer-active-python` (experimental)
 
 **Type**: boolean
